refactor: refactor the tests
