+++ crates/amaru-network/src/tx_submission/tests/mod.rs
@@ -12,17 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod era_tx_id_ord;
-mod server_params;
-mod tx_client_transport;
-mod tx_server_transport;
-mod tx_submission_client;
-mod tx_submission_server;
+mod assertions;
+mod faulty_tx_validator;
+mod message_eq;
+mod mock_transport;
+mod node;
+mod sized_mempool;
+mod test_cases;
+mod tx;
 
-pub use era_tx_id_ord::*;
-pub use server_params::*;
-pub use tx_submission_client::*;
-pub use tx_submission_server::*;
-
-#[cfg(test)]
-mod tests;
+pub use assertions::*;
+pub use faulty_tx_validator::*;
+pub use message_eq::*;
+pub use mock_transport::*;
+pub use node::*;
+pub use sized_mempool::*;
+pub use test_cases::*;
+pub use tx::*;
