+++ crates/amaru-network/src/tx_submission/tx_submission_client.rs
@@ -72,9 +72,9 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
         loop {
             let request = match transport.next_request().await {
                 Ok(r) => r,
-                Err(_) => {
+                Err(e) => {
                     debug!(peer = %self.peer,
-                        "Error receiving next request, terminating tx submission client"
+                        "Error receiving next request, terminating tx submission client: {e:?}"
                     );
                     break;
                 }
@@ -90,7 +90,7 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
                     }
                     if !self
                         .mempool
-                        .wait_for_at_least(self.next_seq().add(required_next as u64))
+                        .wait_for_at_least(self.mempool.last_seq_no().add(required_next as u64))
                         .await
                     {
                         transport.send_done().await?;
@@ -123,17 +123,19 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
                         break;
                     }
                     let txs = self.mempool.get_txs_for_ids(ids.as_slice());
-                    transport
-                        .reply_txs(
-                            txs.into_iter()
-                                .map(|tx| {
-                                    let tx_body = to_cbor(&*tx);
-                                    // TODO: see how to handle multiple eras
-                                    EraTxBody(Era::Conway.into(), tx_body)
-                                })
-                                .collect::<Vec<_>>(),
-                        )
-                        .await?;
+                    if !txs.is_empty() {
+                        transport
+                            .reply_txs(
+                                txs.into_iter()
+                                    .map(|tx| {
+                                        let tx_body = to_cbor(&*tx);
+                                        // TODO: see how to handle multiple eras
+                                        EraTxBody(Era::Conway.into(), tx_body)
+                                    })
+                                    .collect::<Vec<_>>(),
+                            )
+                            .await?;
+                    }
                 }
             }
         }
@@ -191,11 +193,10 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
 pub(crate) mod tests {
     use super::*;
     use crate::tx_submission::tests::{
-        Tx, assert_next_message, assert_tx_bodies_reply, assert_tx_ids_reply, create_transactions,
-        new_era_tx_body, new_era_tx_id,
+        MessageEq, SizedMempool, Tx, assert_next_message, assert_tx_bodies_reply,
+        assert_tx_ids_reply, create_transactions, to_era_tx_bodies, to_era_tx_ids,
     };
     use crate::tx_submission::tx_client_transport::tests::MockClientTransport;
-    use amaru_mempool::strategies::InMemoryMempool;
     use pallas_network::miniprotocols::txsubmission::Message;
     use tokio::sync::mpsc;
     use tokio::sync::mpsc::{Receiver, Sender};
@@ -204,14 +205,15 @@ pub(crate) mod tests {
     #[tokio::test]
     async fn serve_transactions_blocking() -> anyhow::Result<()> {
         // Create a mempool with some transactions
-        let mempool: Arc<InMemoryMempool<Tx>> = Arc::new(InMemoryMempool::default());
-        let txs = create_transactions();
+        let mempool = Arc::new(SizedMempool::default(6));
+        let txs = create_transactions(6);
         let era_tx_ids = to_era_tx_ids(&txs);
         let era_tx_bodies = to_era_tx_bodies(&txs);
         for tx in txs.into_iter() {
             mempool.add(tx)?;
         }
-        let (tx_req, mut replies, _client_handle) = start_client(mempool).await?;
+        let (tx_req, mut replies, _rx_observe_messages, _client_handle) =
+            start_client(mempool).await?;
 
         // Send requests to retrieve transactions and block until they are available.
         // In this case they are immediately available since we pre-populated the mempool.
@@ -244,14 +246,15 @@ pub(crate) mod tests {
     #[tokio::test]
     async fn serve_transactions_non_blocking() -> anyhow::Result<()> {
         // Create a mempool with some transactions
-        let mempool: Arc<InMemoryMempool<Tx>> = Arc::new(InMemoryMempool::default());
-        let txs = create_transactions();
+        let mempool = Arc::new(SizedMempool::default(6));
+        let txs = create_transactions(6);
         let era_tx_ids = to_era_tx_ids(&txs);
         let era_tx_bodies = to_era_tx_bodies(&txs);
         for tx in txs.iter().take(2) {
             mempool.add(tx.clone())?;
         }
-        let (tx_req, mut replies, _client_handle) = start_client(mempool.clone()).await?;
+        let (tx_req, mut replies, _rx_observe_messages, _client_handle) =
+            start_client(mempool.clone()).await?;
 
         // Send requests to retrieve transactions and block until they are available.
         // In this case they are immediately available since we pre-populated the mempool.
@@ -267,7 +270,6 @@ pub(crate) mod tests {
         assert_next_message(&mut replies, Message::Init).await?;
         assert_tx_ids_reply(&mut replies, &era_tx_ids, &[0, 1]).await?;
         assert_tx_bodies_reply(&mut replies, &era_tx_bodies, &[0, 1]).await?;
-        assert_tx_ids_reply(&mut replies, &era_tx_ids, &[]).await?;
 
         // Refill the mempool with more transactions
         for tx in &txs[2..] {
@@ -286,6 +288,7 @@ pub(crate) mod tests {
         // Check replies
         // We basically assert that we receive the expected ids and transactions
         // 2 by 2, then the last one, since we requested batches of 2.
+        assert_tx_ids_reply(&mut replies, &era_tx_ids, &[]).await?;
         assert_tx_ids_reply(&mut replies, &era_tx_ids, &[2, 3]).await?;
         assert_tx_bodies_reply(&mut replies, &era_tx_bodies, &[2, 3]).await?;
         assert_tx_ids_reply(&mut replies, &era_tx_ids, &[4, 5]).await?;
@@ -297,13 +300,14 @@ pub(crate) mod tests {
     #[tokio::test]
     async fn request_txs_must_come_from_requested_ids() -> anyhow::Result<()> {
         // Create a mempool with some transactions
-        let mempool: Arc<InMemoryMempool<Tx>> = Arc::new(InMemoryMempool::default());
-        let txs = create_transactions();
+        let mempool = Arc::new(SizedMempool::default(6));
+        let txs = create_transactions(4);
         let era_tx_ids = to_era_tx_ids(&txs);
         for tx in txs.iter() {
             mempool.add(tx.clone())?;
         }
-        let (tx_req, mut replies, _client_handle) = start_client(mempool.clone()).await?;
+        let (tx_req, mut replies, _rx_observe_messages, _client_handle) =
+            start_client(mempool.clone()).await?;
 
         // Send requests to retrieve transactions and block until they are available.
         // In this case they are immediately available since we pre-populated the mempool.
@@ -324,8 +328,9 @@ pub(crate) mod tests {
 
     #[tokio::test]
     async fn requested_ids_must_be_greater_than_0() -> anyhow::Result<()> {
-        let mempool: Arc<InMemoryMempool<Tx>> = Arc::new(InMemoryMempool::default());
-        let (tx_req, mut replies, _client_handle) = start_client(mempool.clone()).await?;
+        let mempool = Arc::new(SizedMempool::default(6));
+        let (tx_req, mut replies, _rx_observe_messages, _client_handle) =
+            start_client(mempool.clone()).await?;
 
         let requests = vec![Message::RequestTxIds(true, 0, 0)];
         for r in requests {
@@ -339,13 +344,14 @@ pub(crate) mod tests {
 
     #[tokio::test]
     async fn requested_txs_must_be_greater_than_0() -> anyhow::Result<()> {
-        let mempool: Arc<InMemoryMempool<Tx>> = Arc::new(InMemoryMempool::default());
-        let txs = create_transactions();
+        let mempool = Arc::new(SizedMempool::default(4));
+        let txs = create_transactions(4);
         let era_tx_ids = to_era_tx_ids(&txs);
         for tx in txs.iter() {
             mempool.add(tx.clone())?;
         }
-        let (tx_req, mut replies, _client_handle) = start_client(mempool.clone()).await?;
+        let (tx_req, mut replies, _rx_observe_messages, _client_handle) =
+            start_client(mempool.clone()).await?;
 
         let requests = vec![
             Message::RequestTxIds(true, 0, 2),
@@ -363,18 +369,6 @@ pub(crate) mod tests {
 
     // HELPERS
 
-    fn to_era_tx_ids(txs: &[Tx]) -> Vec<EraTxId> {
-        txs.iter()
-            .map(|tx| new_era_tx_id(tx.tx_id()))
-            .collect::<Vec<_>>()
-    }
-
-    fn to_era_tx_bodies(txs: &[Tx]) -> Vec<EraTxBody> {
-        txs.iter()
-            .map(|tx| new_era_tx_body(tx.tx_body()))
-            .collect::<Vec<_>>()
-    }
-
     /// Send a series of requests to a tx submission client backed by the given mempool.
     /// Returns a receiver to collect the messages that the client sent back.
     async fn start_client(
@@ -382,17 +376,19 @@ pub(crate) mod tests {
     ) -> anyhow::Result<(
         Sender<Message<EraTxId, EraTxBody>>,
         Receiver<Message<EraTxId, EraTxBody>>,
+        Receiver<MessageEq>,
         JoinHandle<anyhow::Result<()>>,
     )> {
         let mut client = TxSubmissionClient::new(mempool.clone(), Peer::new("peer-1"));
 
         let (tx_req, rx_req) = mpsc::channel(10);
         let (tx_messages, rx_messages) = mpsc::channel(10);
+        let (tx_observe_messages, rx_observe_messages) = mpsc::channel(10);
 
-        let transport = MockClientTransport::new(rx_req, tx_messages);
+        let transport = MockClientTransport::new(rx_req, tx_observe_messages, tx_messages);
 
         let client_handle =
             tokio::spawn(async move { client.start_client_with_transport(transport).await });
-        Ok((tx_req, rx_messages, client_handle))
+        Ok((tx_req, rx_messages, rx_observe_messages, client_handle))
     }
 }
